Add antibody catpure support
This commit adds support for antibody catpure assays, so they are now
fully supported. It also adjusts the testing script to be agnostic of
absolute path, and doesn't store the test result in the actual repo.
Also, reformatted the `LIB_TYPES_TO_PROGRAM` `dict` in
`scbl_utils/utils/defaults.py` to be more easily legible.