Add dynamic Google Sheet parsing behavior
This commit allows dynamic Google Sheet parsing behavior, so that the
user doesn't need to specify the `header_row` in
`trackingsheet-spec.yml` and `metricssheet-spec.yml`. This also means
that updated tracking sheets or metrics delivery sheets shouldn't break
the script, unless the names of columns change, in which case only the
aforementioned specs need to change.

Additionally, this commit simplifies some code and bumps versions of
dependencies.